+++ backend/daos/mongodbClient.js
@@ -11,8 +11,7 @@ async function connect() {
   }
 
   await mongoose.connect(
-   // `mongodb://${username}:${password}@${address}:27017/`,
-    `mongodb://127.0.0.1:27017/test`,
+    `mongodb://${username}:${password}@${address}:27017/`,
     {
       // mongoose automatically does connection pooling and by default max pool size is 100.
       // We want this to be smaller to reduce load on server.
